refactor: split EmailService into SendEmailService and SendEmailTemplateService
