+++ src/deps.ts
@@ -5,7 +5,8 @@ import { ApiWeatherService, WeatherService } from './services/weather.service.js
 import { SubscriptionService, WeatherSubscriptionService } from './services/subscription.service.js';
 import { Server, HonoServer } from './server.js';
 import { App, CronServerApp } from './app.js';
-import { EmailService, ResendEmailService } from './services/email.service.js';
+import { SendEmailService, ResendSendEmailService } from 'src/services/send-email.service.js';
+import { SendEmailTemplateService, JsxSendEmailTemplateService } from './services/send-email-template.service.js';
 import { LoggerService, PinoLoggerService } from './services/logger.service.js';
 import { DrizzleDbService } from './services/db.service.js';
 
@@ -15,12 +16,18 @@ export const makeDeps = () => {
   const subscriptionRepository: SubscriptionRepository = new DrizzleSubscriptionRepository(dbService.getDb());
   const jwtService: JwtService = new FastJwtService();
   const weatherService: WeatherService = new ApiWeatherService();
-  const emailService: EmailService = new ResendEmailService(loggerService.createLogger('ResendEmailService'));
+  const sendEmailService: SendEmailService = new ResendSendEmailService(
+    loggerService.createLogger('ResendSendEmailService'),
+  );
+  const sendEmailTemplateService: SendEmailTemplateService = new JsxSendEmailTemplateService(
+    sendEmailService,
+    loggerService.createLogger('JsxSendEmailTemplateService'),
+  );
   const subscriptionService: SubscriptionService = new WeatherSubscriptionService(
     jwtService,
     subscriptionRepository,
     weatherService,
-    emailService,
+    sendEmailTemplateService,
     loggerService.createLogger('WeatherSubscriptionService'),
   );
   const cronService: CronService = new CronerCronService(subscriptionService);
@@ -33,7 +40,8 @@ export const makeDeps = () => {
     server,
     cronService,
     subscriptionService,
-    emailService,
+    sendEmailService,
+    sendEmailTemplateService,
     weatherService,
     jwtService,
     subscriptionRepository,
